cleanup:: diff version state, move VersionSelect comp to new file
